+++ claim-db-worker/app/api/analytics/route.ts
@@ -1,12 +1,12 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getEnv } from "@/lib/env";
-import { getClientIP } from "@/lib/utils";
+import { buildRateLimitKey } from "@/lib/server/ratelimit";
 
 export async function POST(request: NextRequest) {
   const env = getEnv();
 
   const url = new URL(request.url);
-  const key = `${getClientIP(request)}:${request.url}`;
+  const key = buildRateLimitKey(request);
 
   // --- Simple rate limiting ---
   const { success } = await env.CLAIM_DB_RATE_LIMITER.limit({ key });
